Factored the GoMath class out of GoMath and made it into GoStoneGroups
