+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "goban",
-    "version": "0.5.118",
+    "version": "0.6.0",
     "description": "",
     "main": "lib/goban.js",
     "types": "lib/goban.d.ts",
