+++ src/__tests__/GoMath.test.ts
@@ -1,8 +1,9 @@
-import { GoMath, BoardState } from "../GoMath";
+import { BoardState } from "../GoStoneGroup";
+import { GoStoneGroups } from "../GoStoneGroups";
 import { JGOFNumericPlayerColor } from "../JGOF";
-import * as GoMathFunction from "../GoMath";
+import * as GoMath from "../GoMath";
 
-describe("GoMath constructor", () => {
+describe("GoStoneGroups constructor", () => {
     test("basic board state", () => {
         const THREExTHREE_board: Array<Array<JGOFNumericPlayerColor>> = [
             [1, 0, 2],
@@ -23,24 +24,24 @@ describe("GoMath constructor", () => {
             foreachNeighbor: () => {},
         };
 
-        const gomath_obj = new GoMath(board_state);
+        const board = new GoStoneGroups(board_state);
 
         // TODO: examine usage in real code and flesh out expectations to reflect that usage
-        expect(gomath_obj.groups.length).toBe(7);
-        expect(gomath_obj.groups[0]).toBe(undefined); // what does this element represent?
-        expect(gomath_obj.groups[1].points).toEqual([{ x: 0, y: 0 }]);
-        expect(gomath_obj.groups[2].points).toEqual([{ x: 1, y: 0 }]);
-        expect(gomath_obj.groups[3].points).toEqual([{ x: 2, y: 0 }]);
-        expect(gomath_obj.groups[4].points).toEqual([
+        expect(board.groups.length).toBe(7);
+        expect(board.groups[0]).toBe(undefined); // what does this element represent?
+        expect(board.groups[1].points).toEqual([{ x: 0, y: 0 }]);
+        expect(board.groups[2].points).toEqual([{ x: 1, y: 0 }]);
+        expect(board.groups[3].points).toEqual([{ x: 2, y: 0 }]);
+        expect(board.groups[4].points).toEqual([
             { x: 0, y: 1 },
             { x: 0, y: 2 },
         ]);
-        expect(gomath_obj.groups[5].points).toEqual([
+        expect(board.groups[5].points).toEqual([
             { x: 1, y: 1 },
             { x: 2, y: 1 },
             { x: 2, y: 2 },
         ]);
-        expect(gomath_obj.groups[6].points).toEqual([{ x: 1, y: 2 }]);
+        expect(board.groups[6].points).toEqual([{ x: 1, y: 2 }]);
     });
 });
 
@@ -516,7 +517,7 @@ describe("sortMoves", () => {
 describe("ojeSequenceToMoves", () => {
     test("bad sequence", () => {
         expect(() => {
-            GoMathFunction.ojeSequenceToMoves("nonsense");
+            GoMath.ojeSequenceToMoves("nonsense");
         }).toThrow("root");
     });
 
@@ -532,6 +533,6 @@ describe("ojeSequenceToMoves", () => {
             ],
         ],
     ])("id of %s", (sequence, id) => {
-        expect(GoMathFunction.ojeSequenceToMoves(sequence)).toStrictEqual(id);
+        expect(GoMath.ojeSequenceToMoves(sequence)).toStrictEqual(id);
     });
 });
